requirements.txt
PyYAML==3.12
boto3==1.4.2
botocore==1.4.81
cffi==1.9.1
contexter==0.1.3
docutils==0.12
jmespath==0.9.0
plumbum==1.6.2
pycparser==2.17
python-dateutil==2.6.0
s3transfer==0.1.9
six==1.10.0
